[refactor] 로또 당첨 결과 반환값 수정

+++ src/main/java/lotto/domain/WinningLotto.java
@@ -20,7 +20,7 @@ public class WinningLotto {
     }
 
     private int getMatchNumberCnt(Lotto lotto) {
-        return (int) lotto.matchCount(winningLotto);
+        return lotto.matchCount(winningLotto);
     }
 
     private boolean getMatchBonusNumber(Lotto lotto) {
